Refactor: python, cts: Move RestartOnebyOne into the _cts module.
No code changes.
